--- web/src/jira/components/JiraAuthLink.tsx
@@ -1,4 +1,0 @@
-import { Link } from "@mui/material";
-import { keijoJiraApiUrl } from "../jiraConfig";
-
-export const JiraAuthLink = <Link href={keijoJiraApiUrl}>Authorize Keijo to use Jira</Link>;
